requirements.txt
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-property==0.2.2
blinker==1.8.2
cached-property==1.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
deprecation==2.1.0
ecdsa==0.19.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.26.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
MarkupSafe==3.0.1
packaging==24.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.21.0
python-dotenv==1.0.1
python-jose==3.3.0
python-keycloak==4.6.3
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
